Update FR parser to use api package
