+++ solution/parser/ecfr-parser/main.go
@@ -26,9 +26,6 @@ import (
 // TIMELIMIT is the total amount of time the process has to run before being cancelled and marked as a failure
 const TIMELIMIT = 5000 * time.Second
 
-// DefaultBaseURL is the default eRegs API URL to use if none is specified
-var DefaultBaseURL = "http://localhost:8000/v3/"
-
 // Functions for easy testing via patching
 var (
 	ParseTitlesFunc        = parseTitles
@@ -40,33 +37,6 @@ var (
 )
 
 var config = &api.ParserConfig{}
-
-func init() {
-	api.BaseURL = os.Getenv("EREGS_API_URL_V3")
-	if api.BaseURL == "" {
-		api.BaseURL = DefaultBaseURL
-	}
-}
-
-func getLogLevel(l string) log.Level {
-	switch l {
-	case "warn":
-		return log.WarnLevel
-	case "fatal":
-		return log.FatalLevel
-	case "error":
-		return log.ErrorLevel
-	case "info":
-		return log.InfoLevel
-	case "debug":
-		return log.DebugLevel
-	case "trace":
-		return log.TraceLevel
-	default:
-		log.Warn("[main] '", config.LogLevel, "' is an invalid log level, defaulting to 'warn'.")
-		return log.WarnLevel
-	}
-}
 
 func parseConfig(c *api.ParserConfig) {
 	parsexml.LogParseErrors = c.LogParseErrors
@@ -81,7 +51,7 @@ func parseConfig(c *api.ParserConfig) {
 		c.Retries = 0
 	}
 
-	log.SetLevel(getLogLevel(c.LogLevel))
+	log.SetLevel(api.GetLogLevel(c.LogLevel))
 }
 
 // Only runs if parser is in a Lambda
