+++ solution/parser/fr-parser/go.mod
@@ -4,6 +4,7 @@ go 1.16
 
 require (
 	github.com/aws/aws-lambda-go v1.30.0
+	github.com/cmsgov/cmcs-eregulations/api v0.0.0-00010101000000-000000000000
 	github.com/cmsgov/cmcs-eregulations/ecfr-parser v0.0.0-00010101000000-000000000000
 	github.com/cmsgov/cmcs-eregulations/network v0.0.0-00010101000000-000000000000
 	github.com/go-test/deep v1.0.8
@@ -13,3 +14,5 @@ require (
 replace github.com/cmsgov/cmcs-eregulations/ecfr-parser => ../ecfr-parser
 
 replace github.com/cmsgov/cmcs-eregulations/network => ../network
+
+replace github.com/cmsgov/cmcs-eregulations/api => ../api
