+++ solution/parser/fr-parser/main.go
@@ -10,7 +10,8 @@ import (
 	"github.com/cmsgov/cmcs-eregulations/fr-parser/fedreg"
 
 	"github.com/cmsgov/cmcs-eregulations/ecfr-parser/ecfr"
-	ecfrEregs "github.com/cmsgov/cmcs-eregulations/ecfr-parser/eregs"
+
+	"github.com/cmsgov/cmcs-eregulations/api"
 
 	"github.com/aws/aws-lambda-go/lambda"
 
@@ -19,18 +20,6 @@ import (
 
 // TIMELIMIT is the total amount of time the process has to run before being cancelled
 const TIMELIMIT = 5000 * time.Second
-
-// DefaultBaseURL is the default eRegs API URL to use if none is specified
-var DefaultBaseURL = "http://localhost:8000/v3/"
-
-func init() {
-	url := os.Getenv("EREGS_API_URL_V3")
-	if url == "" {
-		url = DefaultBaseURL
-	}
-	ecfrEregs.BaseURL = url
-	eregs.BaseURL = url
-}
 
 func lambdaHandler(ctx context.Context) (string, error) {
 	err := start()
@@ -46,32 +35,12 @@ func main() {
 	}
 }
 
-func getLogLevel(l string) log.Level {
-	switch l {
-	case "warn":
-		return log.WarnLevel
-	case "fatal":
-		return log.FatalLevel
-	case "error":
-		return log.ErrorLevel
-	case "info":
-		return log.InfoLevel
-	case "debug":
-		return log.DebugLevel
-	case "trace":
-		return log.TraceLevel
-	default:
-		log.Warn("[main] '", l, "' is an invalid log level, defaulting to 'warn'.")
-		return log.WarnLevel
-	}
-}
-
-var retrieveConfigFunc = ecfrEregs.RetrieveConfig
+var retrieveConfigFunc = api.RetrieveConfig
 
 //lint:ignore U1000 This is required for the tests to work, even if it is not used in this file.
-var getLogLevelFunc = getLogLevel
+var getLogLevelFunc = api.GetLogLevel
 
-func loadConfig() (*ecfrEregs.ParserConfig, error) {
+func loadConfig() (*api.ParserConfig, error) {
 	log.Info("[main] Loading configuration...")
 	config, _, err := retrieveConfigFunc()
 	if err != nil {
@@ -79,14 +48,14 @@ func loadConfig() (*ecfrEregs.ParserConfig, error) {
 	}
 
 	// parse config here
-	log.SetLevel(getLogLevel(config.LogLevel))
+	log.SetLevel(getLogLevelFunc(config.LogLevel))
 
 	return config, nil
 }
 
 var extractSubchapterPartsFunc = ecfr.ExtractSubchapterParts
 
-func getPartsList(ctx context.Context, t *ecfrEregs.TitleConfig) []string {
+func getPartsList(ctx context.Context, t *api.TitleConfig) []string {
 	var parts []string
 	for _, subchapter := range t.Subchapters {
 		subchapterParts, err := extractSubchapterPartsFunc(ctx, t.Title, &ecfr.SubchapterOption{subchapter[0], subchapter[1]})
